إصلاح خطأين جديدين في smart_sync_manager
🔧 الإصلاحات:
✅ إضافة دالة isLimitExceeded() مفقودة في DataUsageManager:140
✅ إصلاح type casting من int إلى double في recordDataUsage:227

🎯 النتيجة: حل جميع أخطاء method definitions وtype compatibility

+++ mobile/lib/services/smart_sync_manager.dart
@@ -224,7 +224,7 @@ class SmartSyncManager {
       // تسجيل استهلاك البيانات
       final backupSize = newBackup.size ?? 0;
       if (backupSize > 0) {
-        await DataUsageManager.instance.recordDataUsage(backupSize);
+        await DataUsageManager.instance.recordDataUsage(backupSize.toDouble());
       }
       
       // تحديد استراتيجية حل التضارب
